witness/comfyui_client/requirements.txt
requests
websockets
pydantic
Pillow 